backend/requirements.txt
blinker==1.9.0
click==8.1.3.1
Flask==3.1.2
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
